fix: move replica from objects to replica
