+++ adapters/clients/replication_test.go
@@ -426,7 +426,7 @@ func TestReplicationCommit(t *testing.T) {
 func TestReplicationFetchObject(t *testing.T) {
 	t.Parallel()
 
-	expected := objects.Replica{
+	expected := replica.Replica{
 		ID: UUID1,
 		Object: &storobj.Object{
 			MarshallerVersion: 1,
@@ -457,7 +457,7 @@ func TestReplicationFetchObject(t *testing.T) {
 
 func TestReplicationFetchObjects(t *testing.T) {
 	t.Parallel()
-	expected := objects.Replicas{
+	expected := replica.Replicas{
 		{
 			ID: UUID1,
 			Object: &storobj.Object{
