+++ usecases/replica/repairer_test.go
@@ -35,7 +35,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 		adds      = additional.Properties{}
 		proj      = search.SelectProperties{}
 		nilObject *storobj.Object
-		emptyItem = objects.Replica{}
+		emptyItem = Replica{}
 	)
 
 	t.Run("GetContentFromDirectRead", func(t *testing.T) {
@@ -43,7 +43,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -72,7 +72,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: objectWithVectors(id, 3, vectors)}
+			item      = Replica{ID: id, Object: objectWithVectors(id, 3, vectors)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 			digestR4  = []types.RepairResponse{{ID: id.String(), UpdateTime: 4, Err: "conflict"}}
@@ -106,8 +106,8 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item2     = objects.Replica{ID: id, Object: object(id, 2)}
-			item3     = objects.Replica{ID: id, Object: object(id, 3)}
+			item2     = Replica{ID: id, Object: object(id, 2)}
+			item3     = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -135,7 +135,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -164,7 +164,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item1     = objects.Replica{ID: id, Object: object(id, 1)}
+			item1     = Replica{ID: id, Object: object(id, 1)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -185,7 +185,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item1     = objects.Replica{ID: id, Object: object(id, 1)}
+			item1     = Replica{ID: id, Object: object(id, 1)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -209,7 +209,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 0, Deleted: false}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3, Deleted: false}}
 		)
@@ -233,7 +233,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: nil, Deleted: true}
+			item      = Replica{ID: id, Object: nil, Deleted: true}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3, Deleted: false}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3, Deleted: false}}
 		)
@@ -251,7 +251,7 @@ func TestRepairerOneWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: nil, LastUpdateTimeUnixMilli: 3, Deleted: true}
+			item      = Replica{ID: id, Object: nil, LastUpdateTimeUnixMilli: 3, Deleted: true}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3, Deleted: true}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3, Deleted: true}}
 		)
@@ -274,7 +274,7 @@ func TestRepairerExistsWithALL(t *testing.T) {
 		ctx       = context.Background()
 		adds      = additional.Properties{}
 		proj      = search.SelectProperties{}
-		emptyItem = objects.Replica{}
+		emptyItem = Replica{}
 	)
 
 	t.Run("ChangedObject", func(t *testing.T) {
@@ -282,7 +282,7 @@ func TestRepairerExistsWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 			digestR4  = []types.RepairResponse{{ID: id.String(), UpdateTime: 4, Err: "conflict"}}
@@ -322,7 +322,7 @@ func TestRepairerExistsWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item3     = objects.Replica{ID: id, Object: object(id, 3)}
+			item3     = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -352,7 +352,7 @@ func TestRepairerExistsWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -412,7 +412,7 @@ func TestRepairerExistsWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item1     = objects.Replica{ID: id, Object: object(id, 1)}
+			item1     = Replica{ID: id, Object: object(id, 1)}
 			digestR1  = []types.RepairResponse{{ID: id.String(), UpdateTime: 1}}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
@@ -437,7 +437,7 @@ func TestRepairerExistsWithALL(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2, Deleted: false}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3, Deleted: false}}
 		)
@@ -492,7 +492,7 @@ func TestRepairerExistsWithConsistencyLevelQuorum(t *testing.T) {
 		ctx       = context.Background()
 		adds      = additional.Properties{}
 		proj      = search.SelectProperties{}
-		emptyItem = objects.Replica{}
+		emptyItem = Replica{}
 	)
 
 	t.Run("ChangedObject", func(t *testing.T) {
@@ -500,7 +500,7 @@ func TestRepairerExistsWithConsistencyLevelQuorum(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 			digestR4  = []types.RepairResponse{{ID: id.String(), UpdateTime: 4, Err: "conflict"}}
@@ -539,7 +539,7 @@ func TestRepairerExistsWithConsistencyLevelQuorum(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes[:2])
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item3     = objects.Replica{ID: id, Object: object(id, 3)}
+			item3     = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -568,7 +568,7 @@ func TestRepairerExistsWithConsistencyLevelQuorum(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes[:2])
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -625,7 +625,7 @@ func TestRepairerExistsWithConsistencyLevelQuorum(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes)
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item1     = objects.Replica{ID: id, Object: object(id, 1)}
+			item1     = Replica{ID: id, Object: object(id, 1)}
 			digestR1  = []types.RepairResponse{{ID: id.String(), UpdateTime: 1}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3}}
 		)
@@ -650,7 +650,7 @@ func TestRepairerExistsWithConsistencyLevelQuorum(t *testing.T) {
 			f         = newFakeFactory(t, "C1", shard, nodes[:2])
 			finder    = f.newFinder("A")
 			digestIDs = []strfmt.UUID{id}
-			item      = objects.Replica{ID: id, Object: object(id, 3)}
+			item      = Replica{ID: id, Object: object(id, 3)}
 			digestR2  = []types.RepairResponse{{ID: id.String(), UpdateTime: 2, Deleted: false}}
 			digestR3  = []types.RepairResponse{{ID: id.String(), UpdateTime: 3, Deleted: false}}
 		)
@@ -709,7 +709,7 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				objectEx(ids[1], 5, shard, "A"),
 				objectEx(ids[2], 6, shard, "A"),
 			}
-			directRe = []objects.Replica{
+			directRe = []Replica{
 				replica(ids[0], 4, false),
 				replica(ids[1], 5, false),
 				replica(ids[2], 6, false),
@@ -808,7 +808,7 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				objectEx(ids[4], 2, shard, "A"),
 			}
 
-			directRe = []objects.Replica{
+			directRe = []Replica{
 				replica(ids[3], 4, false),
 			}
 			digestR2 = []types.RepairResponse{
@@ -825,11 +825,11 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				{ID: ids[3].String(), UpdateTime: 1},
 				{ID: ids[4].String(), UpdateTime: 3}, // latest
 			}
-			directR2 = []objects.Replica{
+			directR2 = []Replica{
 				replica(ids[0], 2, false),
 				replica(ids[1], 2, false),
 			}
-			directR3 = []objects.Replica{
+			directR3 = []Replica{
 				replica(ids[2], 3, false),
 				replica(ids[4], 3, false),
 			}
@@ -1000,7 +1000,7 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				{ID: ids[2].String(), UpdateTime: 1, Err: "conflict"}, // this one
 			}
 
-			directRe = []objects.Replica{
+			directRe = []Replica{
 				replica(ids[0], 4, false),
 				replica(ids[1], 5, false),
 				replica(ids[2], 6, false),
@@ -1093,13 +1093,13 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				{ID: ids[1].String(), UpdateTime: 1},
 				{ID: ids[2].String(), UpdateTime: 4}, // latest
 			}
-			directR2 = []objects.Replica{
+			directR2 = []Replica{
 				replica(ids[1], 3, false),
 			}
-			directR3 = []objects.Replica{
+			directR3 = []Replica{
 				replica(ids[2], 4, false),
 			}
-			directRe = []objects.Replica{
+			directRe = []Replica{
 				replica(ids[0], 2, false),
 				replica(ids[1], 3, false),
 			}
@@ -1186,10 +1186,10 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				{ID: ids[1].String(), UpdateTime: 3},
 				{ID: ids[2].String(), UpdateTime: 4}, // latest
 			}
-			directR2 = []objects.Replica{
+			directR2 = []Replica{
 				replica(ids[1], 3, false),
 			}
-			directR3 = []objects.Replica{
+			directR3 = []Replica{
 				replica(ids[2], 4, false),
 			}
 		)
@@ -1255,13 +1255,13 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				{ID: ids[1].String(), UpdateTime: 3},
 				{ID: ids[2].String(), UpdateTime: 4}, // latest
 			}
-			directR2 = []objects.Replica{
+			directR2 = []Replica{
 				replica(ids[1], 3, false),
 			}
 			// unexpected response UpdateTime  is 3 instead of 4
-			directR3 = []objects.Replica{replica(ids[2], 3, false)}
+			directR3 = []Replica{replica(ids[2], 3, false)}
 
-			directRe = []objects.Replica{
+			directRe = []Replica{
 				replica(ids[0], 2, false),
 				replica(ids[1], 3, false),
 			}
@@ -1336,11 +1336,11 @@ func TestRepairerCheckConsistencyAll(t *testing.T) {
 				{ID: ids[1].String(), UpdateTime: 1},
 				{ID: ids[2].String(), UpdateTime: 4, Deleted: true}, // latest
 			}
-			directR2 = []objects.Replica{
+			directR2 = []Replica{
 				replica(ids[1], 3, false),
 			}
 
-			directRe = []objects.Replica{
+			directRe = []Replica{
 				replica(ids[0], 2, false),
 				replica(ids[1], 3, false),
 			}
@@ -1417,7 +1417,7 @@ func TestRepairerCheckConsistencyQuorum(t *testing.T) {
 			{ID: ids[1].String(), UpdateTime: 5},
 			{ID: ids[2].String(), UpdateTime: 3},
 		}
-		directRe = []objects.Replica{
+		directRe = []Replica{
 			replica(ids[0], 4, false),
 			// replica(ids[1], 5, false),
 			replica(ids[2], 6, false),
